用户下单接口 2022/3/2

+++ src/main/java/net/maxwellclass/online/xdclass/model/entity/VideoOrder.java
@@ -15,7 +15,7 @@ public class VideoOrder {
 
     private Integer id;
     private String outTradeNo;
-    private Long state;
+    private Integer state;
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private Date createTime;
     private Integer totalFee;
